test: separate tests out into different files
